ET-686 ETP's AI fields on classic editor

+++ src/resources/js/rsvp-block.js
@@ -194,11 +194,9 @@ var tribe_tickets_rsvp_block = {
 		return (
 			$.trim( $name.val() ).length &&
 				$.trim( $email.val() ).length &&
-				parseFloat( $qty.val() ) > 0
+				parseInt( $qty.val() ) > 0
 		);
 	};
-
-
 
 	/**
 	 * Handle the form submission
